@ndarocha using floating point arithmetic

+++ temperature_conversion/celcius_to_fahrenheit.c
@@ -2,13 +2,14 @@
 
 int main() 
 {
-  int fahr, celcius;
+  //int fahr, celcius; // integer arithmetic
+  float fahr, celcius;
   int lower = 0, upper = 300, step = 20;
   
   fahr = lower;
   while(fahr <= upper) {
-    celcius = 5*(fahr - 32) / 9;
-    printf("%3d\t%6d\n", fahr, celcius);
+    celcius = 5.0*(fahr - 32.0) / 9.0;
+    printf("%3.0f\t%6.1f\n", fahr, celcius);
     fahr = fahr + step;
     }
   return 0;
